Исправление ошибки инициализации экшенов
