+++ src/components/common/spinner/spinner.module.css
@@ -1,7 +1,7 @@
 .preloader {
     position: fixed;
     left: 0;
-    top: 0;
+    top: 7rem;
     right: 0;
     bottom: 0;
     overflow: hidden;
