+++ src/hooks/use-offers.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '.';
-import { fetchOffersAction } from '../store/api-actions';
 import { getOffersIsLoading } from '../store/main-data/main-data.selectors';
+import { fetchOffersAction } from '../store/main-data/main-data.slice';
 
 export default function useOffers(): boolean {
   const dispatch = useAppDispatch();
