+++ src/index.tsx
@@ -4,8 +4,8 @@ import { Provider } from 'react-redux';
 
 import App from './components/app/app';
 import { store } from './store';
+import { checkAuthAction } from './store/user-process/user-process.slice';
 //import ErrorMessage from './components/error-message/error-message';
-import { checkAuthAction } from './store/api-actions';
 
 store.dispatch(checkAuthAction);
