+++ src/store/main-data/main-data.slice.ts
@@ -1,7 +1,9 @@
-import { createSlice } from '@reduxjs/toolkit';
-import { TMainData } from '../../types/state';
-import { NameSpace } from '../../const';
-import { fetchOffersAction } from '../api-actions';
+import { PayloadAction, createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+import { AppDispatch, State, TMainData } from '../../types/state';
+import { APIRoute, AppRoute, NameSpace } from '../../const';
+import { AxiosInstance } from 'axios';
+import { TOfferPreview } from '../../types/offer-preview';
+import { redirectToRoute } from '../actions';
 
 const initialState: TMainData = {
   activeCity: 'Paris',
@@ -10,12 +12,30 @@ const initialState: TMainData = {
   hasError: false,
 };
 
+export const fetchOffersAction = createAsyncThunk<
+  TOfferPreview[],
+  undefined,
+  {
+    dispatch: AppDispatch;
+    state: State;
+    extra: AxiosInstance;
+  }
+>('cities/fetchOffers', async (_arg, { dispatch, extra: api }) => {
+  const { data } = await api.get<TOfferPreview[]>(APIRoute.Offers);
+  if (!data) {
+    dispatch(redirectToRoute(AppRoute.NotFound));
+  }
+  return data;
+});
+
 export const mainData = createSlice({
   name: NameSpace.Main,
   initialState,
   reducers: {
-    changeCity(state, action) {
-      state.activeCity = action.payload;
+    changeCity(state, action: PayloadAction<string | null>) {
+      if (action.payload && state.activeCity !== action.payload) {
+        state.activeCity = action.payload;
+      }
     },
   },
   extraReducers(builder) {
