+++ src/store/user-process/user-process.slice.ts
@@ -1,11 +1,62 @@
-import { createSlice } from '@reduxjs/toolkit';
-import { AuthorizationStatus, NameSpace } from '../../const';
-import { checkAuthAction, loginAction, logoutAction } from '../api-actions';
-import { TUserProcess } from '../../types/state';
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+import {
+  APIRoute,
+  AppRoute,
+  AuthorizationStatus,
+  NameSpace,
+} from '../../const';
+import { AppDispatch, State, TUserProcess } from '../../types/state';
+import { AxiosInstance } from 'axios';
+import { saveToken, dropToken } from '../../services/token';
+import { TAuthData } from '../../types/auth-data';
+import { TUserData } from '../../types/user-data';
+import { requireAuthorization, redirectToRoute } from '../actions';
 
 const initialState: TUserProcess = {
   authorizationStatus: AuthorizationStatus.Unknown,
 };
+
+export const checkAuthAction = createAsyncThunk<
+  void,
+  undefined,
+  {
+    dispatch: AppDispatch;
+    state: State;
+    extra: AxiosInstance;
+  }
+>('user/checkAuth', async (_arg, { extra: api }) => {
+  await api.get(APIRoute.Login);
+});
+
+export const loginAction = createAsyncThunk<
+  void,
+  TAuthData,
+  {
+    dispatch: AppDispatch;
+    state: State;
+    extra: AxiosInstance;
+  }
+>(
+  'user/login',
+  async ({ login: email, password }, { dispatch, extra: api }) => {
+    const {
+      data: { token },
+    } = await api.post<TUserData>(APIRoute.Login, { email, password });
+    saveToken(token);
+    dispatch(requireAuthorization(AuthorizationStatus.Auth));
+    dispatch(redirectToRoute(AppRoute.Main));
+  }
+);
+
+export const logoutAction = createAsyncThunk<
+  void,
+  undefined,
+  { dispatch: AppDispatch; state: State; extra: AxiosInstance }
+>('user/logout', async (_arg, { dispatch, extra: api }) => {
+  await api.delete(APIRoute.Logout);
+  dropToken();
+  dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
+});
 
 export const userProcess = createSlice({
   name: NameSpace.User,
